- updated shared components to latest versions - moved mixin-tester to src/ so it finds the Vue shim file - misc cleanup - app version = 4.3.1
