+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "business-create-ui",
-  "version": "4.3.0",
+  "version": "4.3.1",
   "private": true,
   "appName": "Create UI",
   "sbcName": "SBC Common Components",
@@ -13,24 +13,24 @@
   },
   "dependencies": {
     "@babel/compat-data": "^7.19.1",
-    "@bcrs-shared-components/breadcrumb": "2.0.9",
-    "@bcrs-shared-components/business-lookup": "1.0.7",
-    "@bcrs-shared-components/certify": "2.0.12",
-    "@bcrs-shared-components/completing-party": "2.0.11",
-    "@bcrs-shared-components/confirm-dialog": "1.1.2",
-    "@bcrs-shared-components/contact-info": "1.1.13",
-    "@bcrs-shared-components/corp-type-module": "1.0.7",
-    "@bcrs-shared-components/court-order-poa": "2.0.13",
-    "@bcrs-shared-components/date-picker": "1.1.23",
-    "@bcrs-shared-components/document-delivery": "1.0.5",
-    "@bcrs-shared-components/effective-date-time": "1.0.21",
-    "@bcrs-shared-components/enums": "1.0.22",
-    "@bcrs-shared-components/help-business-number": "1.0.1",
-    "@bcrs-shared-components/interfaces": "1.0.43",
-    "@bcrs-shared-components/nature-of-business": "1.0.15",
-    "@bcrs-shared-components/staff-comments": "1.2.10",
-    "@bcrs-shared-components/staff-payment": "2.0.15",
-    "@bcrs-shared-components/web-chat": "2.0.1",
+    "@bcrs-shared-components/breadcrumb": "2.1.1",
+    "@bcrs-shared-components/business-lookup": "1.1.1",
+    "@bcrs-shared-components/certify": "2.1.1",
+    "@bcrs-shared-components/completing-party": "2.1.1",
+    "@bcrs-shared-components/confirm-dialog": "1.2.1",
+    "@bcrs-shared-components/contact-info": "1.2.1",
+    "@bcrs-shared-components/corp-type-module": "1.0.8",
+    "@bcrs-shared-components/court-order-poa": "2.1.1",
+    "@bcrs-shared-components/date-picker": "1.2.1",
+    "@bcrs-shared-components/document-delivery": "1.1.1",
+    "@bcrs-shared-components/effective-date-time": "1.1.1",
+    "@bcrs-shared-components/enums": "1.0.23",
+    "@bcrs-shared-components/help-business-number": "1.1.1",
+    "@bcrs-shared-components/interfaces": "1.0.44",
+    "@bcrs-shared-components/nature-of-business": "1.1.1",
+    "@bcrs-shared-components/staff-comments": "1.3.1",
+    "@bcrs-shared-components/staff-payment": "2.1.1",
+    "@bcrs-shared-components/web-chat": "2.1.1",
     "@mdi/font": "^5.9.55",
     "@sentry/browser": "^5.30.0",
     "@sentry/integrations": "^5.30.0",
