+++ tests/unit/date-mixin.spec.ts
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import Vuetify from 'vuetify'
 import { shallowMount } from '@vue/test-utils'
 import { getVuexStore } from '@/store'
-import MixinTester from './mixin-tester.vue'
+import MixinTester from '@/mixin-tester.vue'
 
 Vue.use(Vuetify)
